game-dialog-translator/requirements.txt
streamlit
deepl
python-dotenv
pydantic
pydantic-settings
pytest
requests
